update more features

+++ app/src/main/java/com/teh/windowmodeswitch/QSTileService.kt
@@ -9,25 +9,29 @@ import com.topjohnwu.superuser.Shell
 class QSTileService: TileService() {
     override fun onClick() {
         super.onClick()
-        if (Shell.cmd("getprop ro.boot.slot_suffix").exec().out.contains("_a")) {
-            Log.e("WindowsSwitch", "slot a")
-            Shell.cmd(
-                "dd if=/sdcard/windows/boot.img of=/dev/block/sde14 bs=16M",
-                "sleep 1",
-                "svc power reboot"
-            ).exec()
-        } else if (Shell.cmd("getprop ro.boot.slot_suffix").exec().out.contains("_b")) {
-            Log.e("WindowsSwitch", "slot b")
-            Shell.cmd(
-                "dd if=/sdcard/windows/boot.img of=/dev/block/sde37 bs=16M",
-                "sleep 1",
-                "svc power reboot"
-            ).exec()
+        Shell.getShell()
+        if (Shell.isAppGrantedRoot() == true) {
+            if (Shell.cmd("getprop ro.boot.slot_suffix").exec().out.contains("_a")) {
+                Log.e("WindowsSwitch", "slot a")
+                Shell.cmd(
+                    "dd if=/sdcard/windows/boot.img of=/dev/block/sde14 bs=16M",
+                    "sleep 1",
+                    "svc power reboot"
+                ).exec()
+            } else if (Shell.cmd("getprop ro.boot.slot_suffix").exec().out.contains("_b")) {
+                Log.e("WindowsSwitch", "slot b")
+                Shell.cmd(
+                    "dd if=/sdcard/windows/boot.img of=/dev/block/sde37 bs=16M",
+                    "sleep 1",
+                    "svc power reboot"
+                ).exec()
+            }
         }
 
     }
 
     override fun onTileAdded() {
         super.onTileAdded()
+        Toast.makeText(this, "added qs tile", Toast.LENGTH_SHORT).show()
     }
 }
